Refactor test cases to use testutil.PersonBuilder
